layout: add header with nav bar, search form and login/logout; refactor login/logout display into header
